refactor(validators): document functions and split util functions
