+++ validators/course.go
@@ -1,9 +1,20 @@
+/**
+ * file: validators/course.go
+ * author: theo technciguy
+ * license: apache-2.0
+ *
+ * The course validator validates the course field of the
+ * feedback form.
+ */
+
+// Package validators provides validation functions for the
+// feedback form.
+// It also includes custom validators for specific fields.
 package validators
 
 import (
 	"fmt"
 	"net/http"
-	"reflect"
 	"strconv"
 	"strings"
 
@@ -46,6 +57,9 @@ func IsCourse(fl validator.FieldLevel) bool {
 	return true
 }
 
+// ValidateFeedback validates the feedback structure. It returns an
+// UXErrors containing all the errors that occurred during validation
+// or nil if no errors occurred.
 func ValidateFeedback(f *models.Feedback) error {
 	v := validator.New()
 	v.RegisterValidation("iscourse", IsCourse, false)
@@ -81,46 +95,4 @@ func ValidateFeedback(f *models.Feedback) error {
 	}
 
 	return errs
-}
-
-func requiredMissingError(xerr *uxerrors.Error, err validator.FieldError) {
-	xerr.Summary = fmt.Sprintf("The %s field is missing", err.Field())
-	xerr.Detail = fmt.Sprintf("The %s field is a required field, however it is empty. Fill the field correctly and try again.", err.Field())
-}
-
-func genericError(xerr *uxerrors.Error, err validator.FieldError) {
-	xerr.Summary = fmt.Sprintf("There was an error while validating the %s field", err.Field())
-	xerr.Detail = fmt.Sprintf("While validating the %s field, an unspecified error occurred. Check the field for correctness and try again.", err.Field())
-}
-
-func minError(xerr *uxerrors.Error, err validator.FieldError) {
-	var summary, detail string
-
-	switch err.Type() {
-	case reflect.TypeFor[string]():
-		summary = "is too short"
-		detail = fmt.Sprintf("It should be at least %s long, but was %d. Elaborate and try again.", err.Param(), len(err.Value().(string)))
-	case reflect.TypeFor[int](), reflect.TypeFor[uint]():
-		summary = "is too small"
-		detail = fmt.Sprintf("It should be at least %s, but was %d. Increase the value and try again.", err.Param(), err.Value())
-	}
-
-	xerr.Summary = fmt.Sprintf("The %s field %s", err.Field(), summary)
-	xerr.Detail = fmt.Sprintf("The %s field %s. %s", err.Field(), summary, detail)
-}
-
-func maxError(xerr *uxerrors.Error, err validator.FieldError) {
-	var summary, detail string
-
-	switch err.Type() {
-	case reflect.TypeFor[string]():
-		summary = "is too long"
-		detail = fmt.Sprintf("It should be at most %s long, but was %d. Shorten and try again.", err.Param(), len(err.Value().(string)))
-	case reflect.TypeFor[int](), reflect.TypeFor[uint]():
-		summary = "is too small"
-		detail = fmt.Sprintf("It should be at most %s, but was %d. Decrease the value and try again.", err.Param(), err.Value())
-	}
-
-	xerr.Summary = fmt.Sprintf("The %s field %s", err.Field(), summary)
-	xerr.Detail = fmt.Sprintf("The %s field %s. %s", err.Field(), summary, detail)
 }
